+++ src/scenes/implementations/buttons.cpp
@@ -16,7 +16,7 @@ Buttons::Buttons(Scenes &scenes) : Scene(scenes) {
             {{0.5f, 0.5f, 0.0f}, {0.0f, 0.0f, 1.0f, 1.0f}, {1.0f, 1.0f}},
             {{-0.5f, 0.5f, 0.0f}, {1.0f, 1.0f, 1.0f, 1.0f}, {0.0f, 1.0f}},
         },
-        std::vector<uint16_t> {
+        std::vector<uint32_t> {
             0, 1, 2, 2, 3, 0,
         }
     );
